Add an optional preview window
Consider this to be an experiment. Perhaps we could extend this further
to show progress bars, allow live re-rendering on scene changes, etc.

Baseline is a window that pops up and just renders the pixel data as-is
while the image is being built.

This slows down rendering by quite a large factor right now, but I think
it can be optimized a bit. We'd need some profiling to determine what is
the problem with that, but I suspect all the locking while rendering is
to blame.