Improve preview window and reduce runtime tax
1. Make window prettier and prevent resizing
2. Separate texture updating, rendering and close conditions from each
   other.

The window.next() method returns a lot of events. When mousing over the
screen, when it's time to render, when it's time to update, and so on.

We can adjust how often these events come in and then only update the
texture when asked to update and render the already done texture when
asked to render. This speeds up the render pipeline quite a bit!

Also note how the FPS (render) and UPS (update) have been adjusted to
sane values.