Switch {Mutex,RwLock} for parking_lot versions
parking_lot is a lot(!) faster so this removed almost all of the
remaining performance taxes.

It works by a combination of multiple things:
  1. Remove Mutex poisoning (e.g. a panic while holding the lock just
     unlocks)

  2. Removes heap indirection for Mutex (e.g. it's now Sized and on the
     stack)

  3. parking_lot::Mutex has optimized Mutex a lot for uncontested and
     for very short locks, which is what we have here.

All of this is combined into a nice package that gives us much faster
performance.

My example run went from 1.2 seconds to 600ms, so it's double speed now.
Strangely enough, it's also consistently *faster* with the preview
window than without it. I have no idea why as it does not make any
sense.